requirements.txt
langchain
openai
streamlit
PyPDF2 
langchain-openai 
tiktoken
chromadb
python-dotenv
fpdf